requirements.txt
langchain
langgraph
langchain-google-genai
langchain-groq
python-dotenv
numpy
termcolor
streamlit-plotly-events
plotly